More -Wconversion fixes in examples

+++ examples/fem_system/fem_system_ex4/heatsystem.C
@@ -80,7 +80,7 @@ bool HeatSystem::element_time_derivative (bool request_jacobian,
 
   // First we get some references to cell-specific data that
   // will be used to assemble the linear system.
-  const unsigned int dim = c.get_elem().dim();
+  const unsigned short dim = c.get_elem().dim();
   FEBase * fe = nullptr;
   c.get_element_fe(T_var, fe, dim);
